waveform: Move ExtendedPropertyValue type alias to nitypes.waveform.typing
